+++ src/main/java/lotto/domain/LottoNumbers.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
-import lotto.generator.NumberGenerateStrategy;
+import lotto.domain.generator.NumberGenerateStrategy;
 
 public class LottoNumbers {
